backend/requirements.txt
fastapi
motor
python-dotenv
uvicorn[standard]

